Users determine the number of posts to display and UI enhancements
Users can set the number of headlines they want to issue per day. This is useful when the plugin is installed on different websites and business decisions want to publish a number than the default 5 posts per day.

Cleaned up the Dashboard widget so that instructions for the widget exist behind a button instead of always open and taking up space.